Move views into Views directory and namespace
