+++ tools/PI/DevHome.PI/Views/BarWindow.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using DevHome.Common.Extensions;
-using DevHome.PI.Helpers;
 using DevHome.PI.Models;
 using DevHome.PI.Properties;
 using DevHome.PI.ViewModels;
@@ -15,7 +14,7 @@ using Windows.Graphics;
 using Windows.Win32.Foundation;
 using WinUIEx;
 
-namespace DevHome.PI;
+namespace DevHome.PI.Views;
 
 public partial class BarWindow
 {
