+++ tools/PI/DevHome.PI/Views/BarWindowHorizontal.xaml.cs
