+++ tools/PI/DevHome.PI/Views/BarWindowVertical.xaml
